Fix redirectUrl
