+++ lib/modules/arguments-form-widgets/views/widget.html
@@ -3,7 +3,7 @@
 		<form
 			method="POST"
 			action="/modules/arguments-form-widgets/submit"
-			class="argument-form clearfix {{ 'logged-in' if data.loggedIn else 'not-logged-in' }}"
+			class="argument-form clearfix ajax-refresh-after-submit {{ 'logged-in' if data.loggedIn else 'not-logged-in' }} "
 		>
 			<div class="form-group">
 				<textarea
