+++ lib/modules/arguments-widgets/views/reaction-form.html
@@ -1,4 +1,4 @@
-<form method="POST" action="{{apiArgumentUrl}}" class="newReaction argument-form reply-form ">
+<form method="POST" action="{{apiArgumentUrl}}" class="newReaction argument-form reply-form ajax-form ajax-refresh-after-submit">
   <textarea name="description" rows="2" placeholder="Typ hier jouw reactie..."></textarea>
 
   <input type="hidden" name="parentId" value="{{arg.id}}">
