+++ lib/modules/idea-pages/index.js
@@ -1,4 +1,6 @@
 const rp  = require('request-promise');
+const Url = require('url');
+const eventEmitter  = require('../../../events').emitter;
 
 module.exports = {
   extend: 'apostrophe-custom-pages',
@@ -70,6 +72,88 @@ module.exports = {
           callback(null);
         });
     });
+
+
+
+   self.apos.app.get('/like', (req, res, next) => {
+     if (
+       req.data.global.siteConfig && req.data.global.siteConfig.votes
+       && req.data.global.siteConfig.votes.voteType !== 'likes'
+     ) {
+       throw Error('GET Route only allowed for vote type like');
+     }
+
+     req.body.votes = [{
+       ideaId: req.query.ideaId,
+       opinion: req.query.opinion ? req.query.opinion : 'yes',
+     }];
+
+     req.redirectUrl = req.query.redirectUrl ? req.query.redirectUrl : '/' + req.data.global.ideaSlug + '/' + req.query.ideaId;
+
+     postVote(req, res, next);
+   });
+
+   const postVote = (req, res, next) => {
+    eventEmitter.emit('vote');
+
+     const apiUrl = self.apos.settings.getOption(req, 'apiUrl');
+     const siteId = req.data.global.siteId;
+     const postUrl = `${apiUrl}/api/site/${siteId}/vote`;
+     const ip = req.headers['x-forwarded-for'] || req.connection.remoteAddress;
+
+     let votes = req.body.votes ? req.body.votes : [{
+       ideaId: req.body.ideaId,
+       opinion: req.body.opinion,
+    //   ipOriginXXX: ip // 1111
+     }];
+
+     const options = {
+         method: 'POST',
+         uri: postUrl,
+         headers: {
+             'Accept': 'application/json',
+             "X-Authorization" : `Bearer ${req.session.jwt}`,
+         },
+         body: votes,
+         json: true // Automatically parses the JSON string in the response
+     };
+
+     rp(options)
+      .then(function (response) {
+        if (req.redirectUrl) {
+          var redirectUrl = Url.parse(req.redirectUrl);
+          redirectUrl = redirectUrl ? redirectUrl.pathname : '/';
+          res.redirect(redirectUrl);
+        } else {
+          res.end(JSON.stringify({
+            id: response.id
+          }));
+        }
+      })
+      .catch(function (err) {
+          console.log(err);
+
+          res.status(500).json(err);
+       });
+   }
+
+
+   self.apos.app.post('/vote', (req, res, next) => {
+     postVote(req, res, next);
+   });
+
+   const superPageBeforeSend = self.pageBeforeSend;
+
+   self.pageBeforeSend = (req, callback) => {
+     const pageData = req.data.page;
+
+     if (req.query.voteOpinion && req.query.ideaId) {
+       res.redirect(`/like?ideaId=${req.query.ideaId}&opinion=${req.query.voteOpinion}&redirectUrl=${req.data.currentPath}`)
+     }
+
+     callback();
+     return superPageBeforeSend(req, callback);
+   };
 
   }
 };
