+++ lib/modules/idea-single-widgets/index.js
@@ -13,51 +13,6 @@ module.exports = {
 
      require('./lib/routes.js')(self, options);
 
-     const postVote = (req, res, next) => {
-      eventEmitter.emit('vote');
-
-       const apiUrl = self.apos.settings.getOption(req, 'apiUrl');
-       const siteId = req.data.global.siteId;
-       const postUrl = `${apiUrl}/api/site/${siteId}/vote`;
-       const ip = req.headers['x-forwarded-for'] || req.connection.remoteAddress;
-
-       let votes = req.body.votes ? req.body.votes : [{
-         ideaId: req.body.ideaId,
-         opinion: req.body.opinion,
-      //   ipOriginXXX: ip // 1111
-       }];
-
-       votes = votes.map((vote) => {
-         vote.ipOriginXXX = ip;
-         return vote;
-       })
-
-       const options = {
-           method: 'POST',
-           uri: postUrl,
-           headers: {
-               'Accept': 'application/json',
-               "X-Authorization" : `Bearer ${req.session.jwt}`,
-           },
-           body: votes,
-           json: true // Automatically parses the JSON string in the response
-       };
-
-       rp(options)
-        .then(function (response) {
-          if (req.redirectUrl) {
-            res.redirect(req.redirectUrl);
-          } else {
-            res.end(JSON.stringify({
-              id: response.id
-            }));
-          }
-        })
-        .catch(function (err) {
-            res.status(500).json(err);
-         });
-     }
-
      const superPushAssets = self.pushAssets;
      self.pushAssets = function () {
        superPushAssets();
@@ -115,41 +70,5 @@ module.exports = {
        return superOutput(widget, options);
      };
 
-     self.apos.app.get('/like', (req, res, next) => {
-       if (
-         req.data.global.siteConfig && req.data.global.siteConfig.votes
-         && req.data.global.siteConfig.votes.voteType !== 'likes'
-       ) {
-         throw Error('GET Route only allowed for vote type like');
-       }
-
-       req.body.votes = [{
-         ideaId: req.query.ideaId,
-         opinion: req.query.opinion ? req.query.opinion : 'yes',
-       }];
-
-       req.redirectUrl = '/' + req.data.global.ideaSlug + '/' + req.query.ideaId;
-
-       postVote(req, res, next);
-     });
-
-     self.apos.app.post('/vote', (req, res, next) => {
-       postVote(req, res, next);
-     });
-
-     const superPageBeforeSend = self.pageBeforeSend;
-
-     self.pageBeforeSend = (req, callback) => {
-       const pageData = req.data.page;
-
-       if (req.query.voteOpinion) {
-         req.data.formToSubmit = {
-           'opinion': req.query.voteOpinion
-         }
-       }
-
-       callback();
-       return superPageBeforeSend(req, callback);
-     };
   }
 };
