+++ lib/modules/resource-overview-widgets/views/display/minimalVotes.html
@@ -12,12 +12,11 @@
   data-ideaid="{{idea.id}}"
 >
   <a href="{{data.widget.pathForResource}}/{{idea.id}}">
-
     {% if data.widget.displayRanking and idea.status != 'DENIED' and idea.ranking %}
     <div class="ranking"><div class="label">{{idea.ranking}}</div></div>
     {% endif %}
 
-    <div class="image"   style="background-image: url('{{data.widget.formatImageUrl(idea.extraData.images[0], idea.location, data.widget, 500, 500, true, data.cookieConsent) }}')"></div>
+    <div class="image" style="background-image: url('{{data.widget.formatImageUrl(idea.extraData.images[0], idea.location, data.widget, 500, 500, true, data.cookieConsent) }}')"></div>
 
     <div class="duration">
       {{ data.widget['label' + idea.status] }}
@@ -47,9 +46,11 @@
               <form method="GET" action="/like" class="ajax-form ajax-refresh-after-submit {{'voted' if idea.userVote.opinion == 'yes'}}">
                 <input type="hidden" value="{{idea.id}}" name="ideaId" />
                 <button
-                  data-return-to="{{data.currentPathname}}?voteOpinion=yes"
-                  type="submit" class="count yes role-required-anonymous">
-                    {{idea.yes}}
+                  data-return-to="/like?voteOpinion=yes&ideaId={{idea.id}}&redirectUrl={{data.currentPathname}}"
+                  type="submit"
+                  class="count yes role-required-{{data.global.roleToLike if data.global.roleToLike else 'anonymous'}}"
+                >
+                  {{idea.yes}}
                 </button>
               </form>
               {% else %}
