+++ locales/default.json
@@ -390,5 +390,67 @@
 	"Edit Like": "Edit Like",
 	"Edit Share widgets": "Edit Share widgets",
 	"Edit Location": "Edit Location",
-	"Edit Idea form": "Edit Idea form"
+	"Edit Idea form": "Edit Idea form",
+	"Column Section": "Column Section",
+	"Save Column Section": "Save Column Section",
+	"Background color": "Background color",
+	"Background image": "Background image",
+	"HTML Container id (must be unique on the page for css)": "HTML Container id (must be unique on the page for css)",
+	"Columns": "Columns",
+	"Full page width ": "Full page width ",
+	"One column: 100%": "One column: 100%",
+	"Two Columns: 50% - 50%": "Two Columns: 50% - 50%",
+	"Two Columns: 33% - 66%": "Two Columns: 33% - 66%",
+	"Two Columns: 66% - 33%": "Two Columns: 66% - 33%",
+	"Two Columns: 75% - 25%": "Two Columns: 75% - 25%",
+	"Two Columns: 25% - 75%": "Two Columns: 25% - 75%",
+	"Two Columns: Desktop: 75% - 25%, Tablet:  66% - 33%": "Two Columns: Desktop: 75% - 25%, Tablet:  66% - 33%",
+	"Three Columns: 25% - 50% - 25%": "Three Columns: 25% - 50% - 25%",
+	"Three columns: 33% - 33% - 33%": "Three columns: 33% - 33% - 33%",
+	"Four Columns: 25% - 25% - 25% - 25%": "Four Columns: 25% - 25% - 25% - 25%",
+	"Full screen (vertical & horizontal)": "Full screen (vertical & horizontal)",
+	"Margin type": "Margin type",
+	"Normal": "Normal",
+	"Shift upwards": "Shift upwards",
+	"HTML ID": "HTML ID",
+	"HTML Class": "HTML Class",
+	"Show section as toggle section": "Show section as toggle section",
+	"Toggle title": "Toggle title",
+	"Toggle arrow color": "Toggle arrow color",
+	"White": "White",
+	"Black": "Black",
+	"Section open by default?": "Section open by default?",
+	"Show toggle only on mobile": "Show toggle only on mobile",
+	"Change the settings of a page.": "Change the settings of a page.",
+	"Basics": "Basics",
+	"Permissions": "Permissions",
+	"Slug": "Slug",
+	"Type": "Type",
+	"Resource": "Resource",
+	"Home": "Home",
+	"Published": "Published",
+	"Trash": "Trash",
+	"Hide in Navigation": "Hide in Navigation",
+	"Who can view this?": "Who can view this?",
+	"Public": "Public",
+	"Login Required": "Login Required",
+	"Certain People": "Certain People",
+	"These Users can View": "These Users can View",
+	"These Groups can View": "These Groups can View",
+	"These Users can Edit": "These Users can Edit",
+	"These Groups can Edit": "These Groups can Edit",
+	"Apply permissions to subpages now": "Apply permissions to subpages now",
+	"This is a one-time operation that takes place when you click save.": "This is a one-time operation that takes place when you click save.",
+	"How would you like the permissions to be applied to the subpages ?": "How would you like the permissions to be applied to the subpages ?",
+	"Copy (Overrides the existing permissions)": "Copy (Overrides the existing permissions)",
+	"Append (Adds to the existing permissions)": "Append (Adds to the existing permissions)",
+	"Meta title": "Meta title",
+	"Meta Description": "Meta Description",
+	"Meta Tags": "Meta Tags",
+	"Redirect de gebruiker wanneer niet ingelogd": "Redirect de gebruiker wanneer niet ingelogd",
+	"Hide the footer?": "Hide the footer?",
+	"Search...": "Search...",
+	"Browse": "Browse",
+	"editor2": "editor2",
+	"editor": "editor"
 }
